+++ packages/providers/src/provider.test.ts
@@ -8,6 +8,10 @@ import * as GraphQL from 'graphql-request';
 
 import Provider from './provider';
 
+afterEach(() => {
+  jest.restoreAllMocks();
+});
+
 /*
  * @group common/e2e
  */
@@ -17,7 +21,7 @@ describe('Provider', () => {
 
     const version = await provider.getVersion();
 
-    expect(version).toEqual('0.15.1');
+    expect(version).toEqual('0.17.3');
   });
 
   it('can call()', async () => {
@@ -59,7 +63,7 @@ describe('Provider', () => {
       {
         type: ReceiptType.ScriptResult,
         result: bn(0),
-        gasUsed: bn(0x86b),
+        gasUsed: bn(0x67),
       },
     ];
 
@@ -157,7 +161,7 @@ describe('Provider', () => {
     expect(minGasPrice).toBeDefined();
   });
 
-  it('can change the provider url of the curernt instance', async () => {
+  it('can change the provider url of the current instance', async () => {
     const providerUrl1 = 'http://127.0.0.1:4000/graphql';
     const providerUrl2 = 'http://127.0.0.1:8080/graphql';
     const provider = new Provider(providerUrl1);
@@ -166,6 +170,34 @@ describe('Provider', () => {
     expect(provider.url).toBe(providerUrl1);
     provider.connect(providerUrl2);
     expect(provider.url).toBe(providerUrl2);
-    expect(spyGraphQLClient).toBeCalledWith(providerUrl2);
+    expect(spyGraphQLClient).toBeCalledWith(providerUrl2, undefined);
+  });
+
+  it('can accept a custom fetch function', async () => {
+    const providerUrl = 'http://127.0.0.1:4000/graphql';
+
+    const customFetch = async (
+      url: string,
+      options: {
+        body: string;
+        headers: { [key: string]: string };
+        [key: string]: unknown;
+      }
+    ) => {
+      const graphqlRequest = JSON.parse(options.body);
+      const { operationName } = graphqlRequest;
+      if (operationName === 'getVersion') {
+        const responseText = JSON.stringify({
+          data: { nodeInfo: { nodeVersion: '0.30.0' } },
+        });
+        const response = new Response(responseText, options);
+
+        return response;
+      }
+      return fetch(url, options);
+    };
+
+    const provider = new Provider(providerUrl, { fetch: customFetch });
+    expect(await provider.getVersion()).toEqual('0.30.0');
   });
 });