+++ packages/hdwallet/CHANGELOG.md
@@ -1,5 +1,41 @@
 # Change Log
 
+## 0.33.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fuel-ts/signer@0.33.0
+  - @fuel-ts/math@0.33.0
+  - @fuel-ts/mnemonic@0.33.0
+
+## 0.32.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fuel-ts/signer@0.32.0
+  - @fuel-ts/math@0.32.0
+  - @fuel-ts/mnemonic@0.32.0
+
+## 0.31.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fuel-ts/signer@0.31.0
+  - @fuel-ts/math@0.31.0
+  - @fuel-ts/mnemonic@0.31.0
+
+## 0.30.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @fuel-ts/signer@0.30.0
+  - @fuel-ts/math@0.30.0
+  - @fuel-ts/mnemonic@0.30.0
+
 ## 0.29.1
 
 ### Patch Changes